backend/requirements.txt
Django==3.0.7
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework
django-cors-headers
djangorestframework-simplejwt
